fixup pov-distribution
