+++ node/network/pov-distribution/Cargo.toml
@@ -6,19 +6,13 @@ edition = "2018"
 
 [dependencies]
 futures = "0.3.5"
-log = "0.4.8"
-futures-timer = "3.0.2"
-streamunordered = "0.5.1"
+log = "0.4.11"
 polkadot-primitives = { path = "../../../primitives" }
-node-primitives = { package = "polkadot-node-primitives", path = "../../primitives" }
-parity-scale-codec = "1.3.4"
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master" }
 polkadot-subsystem = { package = "polkadot-node-subsystem", path = "../../subsystem" }
 polkadot-node-subsystem-util = { path = "../../subsystem-util" }
 polkadot-node-network-protocol = { path = "../../network/protocol" }
 
 [dev-dependencies]
-parking_lot = "0.10.0"
 assert_matches = "1.3.0"
 sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 polkadot-node-subsystem-test-helpers = { path = "../../subsystem-test-helpers" }
